Change the type of 'gain' from unsigned short to short
* Modify every printf to fprintf
* Add easter eggs

+++ inc/player.h
@@ -17,8 +17,8 @@
  */
 struct Player_s {
     char *name;          ///< Name of the player
-    unsigned short credits;                ///< Player's credits
-    unsigned short gain;                   ///< Player's gain
+    unsigned int credits;                ///< Player's credits
+    short gain;                   ///< Player's gain
     unsigned char bet;                    ///< Player's bet on the current round
 };
